add errors to purchases

+++ purchases/src/app/purchases/purchases.controller.ts
@@ -1,71 +1,68 @@
-import { Request, Response } from "express";
-import { IPurchasesService } from "./interfaces";
+import { NextFunction, Request, Response } from "express";
+import { IPurchasesController, IPurchasesService } from "./interfaces";
 import { CreatePurchaseDto } from "./dtos";
-
-interface IPurchasesController {
-  findByUserId(req: Request, res: Response): Promise<void>;
-  findById(req: Request, res: Response): Promise<void>;
-  create(req: Request, res: Response): Promise<void>;
-  delete(req: Request, res: Response): Promise<void>;
-}
+import { BadRequestError } from "../../shared/errors";
 
 export class PurchasesController implements IPurchasesController {
   constructor(private _purchasesService: IPurchasesService) {}
 
-  async findByUserId(req: Request, res: Response) {
-    const userId = Number(req.params.userId);
+  async findByUserId(req: Request, res: Response, next: NextFunction) {
+    try {
+      const userId = Number(req.params.userId);
 
-    const response = await this._purchasesService.findByUserId(userId);
+      if (isNaN(userId)) {
+        throw new BadRequestError("Invalid user id");
+      }
 
-    if (response?.message === "The user has no purchases") {
-      res.status(404).send(response);
-      return;
+      const response = await this._purchasesService.findByUserId(userId);
+
+      res.status(200).send(response);
+    } catch (error) {
+      next(error);
     }
-
-    res.status(200).send(response);
   }
 
-  async findById(req: Request, res: Response) {
-    const purchaseId = Number(req.params.purchaseId);
+  async findById(req: Request, res: Response, next: NextFunction) {
+    try {
+      const purchaseId = Number(req.params.purchaseId);
 
-    const response = await this._purchasesService.findById(purchaseId);
+      if (isNaN(purchaseId)) {
+        throw new BadRequestError("Invalid purchase id");
+      }
 
-    if (response?.message === "Purchase not found") {
-      res.status(404).send(response);
-      return;
+      const response = await this._purchasesService.findById(purchaseId);
+
+      res.status(200).send(response);
+    } catch (error) {
+      next(error);
     }
-
-    res.status(200).send(response);
   }
 
-  async create(req: Request, res: Response) {
-    const dto = req.body as CreatePurchaseDto;
+  async create(req: Request, res: Response, next: NextFunction) {
+    try {
+      const dto = req.body as CreatePurchaseDto;
 
-    const response = await this._purchasesService.create(dto);
+      const response = await this._purchasesService.create(dto);
 
-    if (response?.message === "Internal server error") {
-      res.status(500).send(response);
-      return;
+      res.status(201).send(response);
+    } catch (error) {
+      next(error);
     }
-
-    if (response?.message) {
-      res.status(400).send(response);
-      return;
-    }
-
-    res.status(201).send(response);
   }
 
-  async delete(req: Request, res: Response) {
-    const purchaseId = Number(req.params.purchaseId);
+  async delete(req: Request, res: Response, next: NextFunction) {
+    try {
+      const purchaseId = Number(req.params.purchaseId);
 
-    const response = await this._purchasesService.delete(purchaseId);
+      if (isNaN(purchaseId)) {
+        throw new BadRequestError("Invalid purchase id");
+      }
 
-    if (response?.message === "Purchase not found") {
-      res.status(404).send(response);
-      return;
+      await this._purchasesService.delete(purchaseId);
+
+      res.status(204).send();
+    } catch (error) {
+      next(error);
     }
-
-    res.status(204).send();
   }
 }
